backend/tasks/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
python-dateutil==2.8.2
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0